prod/Arm32Edgesolution/ArmEdgeSolution/modules/ArmPredictionModule/requirements.txt
azure-iot-device~=2.7.0
Pillow
pyfiglet